pyproject.toml
[tool.poetry]
name = "baseball crawler"
version = "0.1.0"
description = ""
authors = ["Sean12344321 <Sean930110@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
selenium = "^4.25.0"
mysql-connector-python = "^9.1.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
